requirements.txt
python-osc
pyserial
pyyaml
